Backend/requirements.txt
playwright>=1.41.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.1
openai>=1.12.0
# MCP support (Playwright MCP server integration)
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
Pillow>=10.0.0
fastapi==0.104.0
starlette==0.27.0
uvicorn>=0.24.0

# Persona Generator - LLM Providers
langchain-groq>=0.1.0
langchain-google-genai>=0.0.1

# Web scraping & content extraction
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web search (for persona generation pipeline v2)
ddgs>=6.0
